backend/requirements.txt
email-validator==1.1.2
fastapi==0.62.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3
pydantic~=1.9.0
pymongo~=4.1.1
pytest~=7.1.1
python-decouple==3.3